pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "claudechic"
dynamic = ["version"]
description = "Claude Chic - A stylish terminal UI for Claude Code"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.9.0",
    "anthropic>=0.75.0",
    "claude-agent-sdk>=0.1.40",  # 0.1.40 fixes rate_limit_event crash, enables Opus 4.6
    "psutil>=5.9.0",
    "pyperclip>=1.11.0",
    "pyyaml>=6.0",
    "textual>=7.4.0",
    "textual-autocomplete>=4.0.6",
]
authors = [{ name = "Matthew Rocklin", email = "mrocklin@gmail.com" }]
license = "MIT"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
Homepage = "https://github.com/sprustonlab/claudechic"
Repository = "https://github.com/sprustonlab/claudechic"

[project.scripts]
claudechic = "claudechic.__main__:main"

[dependency-groups]
dev = [
    "altair>=6.0.0",
    "marimo>=0.19.6",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "pre-commit>=4.0.0",
    "pyright>=1.1.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-xdist>=3.8.0",
]

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["claudechic*"]

[tool.setuptools.package-data]
claudechic = [
    "*.tcss",
    "*.md",
    "prompts/*.md",
    "global/*.yaml",
    "context/*.md",
    "workflows/**/*.yaml",
    "workflows/**/*.md",
    "mcp_tools/*.py",
    "mcp_tools/*.yaml",
    "mcp_tools/*.md",
    "audit/*.py",
    "audit/*.md",
]

[tool.pyright]
exclude = ["build", ".venv", "dist", "site", "**/__pycache__"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "fast: fast unit tests (default tier, under 1s each)",
    "integration: integration tests (may be slower, uses more complex mocking)",
    "slow: tests that take >30s (real timeouts, E2E flows)",
]

[tool.setuptools_scm]
